Le retour de la somme au carre est installe, en croisant les doigts que la reponse est juste

+++ RayTracingProject/SourceAndHeadersOnly/coefficients.c
@@ -15,7 +15,7 @@ const float eps_0 = 8.854*pow(10, -12);
 const float omega = (2.45*pow(10,9))*2*pi; //pulsation en rad/s, définie par la fréquence de travail
 const float beta = (2.45*pow(10,9))*2*pi / (3*pow(10,8));
 
-float theta_i(float x0, float y0, float x1, float y1, char vertical){
+double theta_i(float x0, float y0, float x1, float y1, char vertical){
 // Calcul de l'angle d'incidence depuis le "point 0 emetteur" jusqu'au point 1 d'intersection avec le mur
     if( vertical ){
         return fabs( atan( (y1-y0)/(x1-x0) ) );
@@ -46,16 +46,16 @@ float complex gamma_m(float eps_rel, float sigma){
     return alpha + I * beta;
 }
 
-float theta_t_snell(float theta_i, float eps1, float eps2){
+double theta_t_snell(double theta_i, float eps1, float eps2){
     // Calcul de l'angle de transmission par Snell, besoin pour calcul du coefficient de reflexion perpendiculaire
 
     return asin( sqrt(eps1/eps2)*sin(theta_i) );
 }
 
-float norme_coeff_reflexion(float theta_i, float eps_rel, float sigma, float epaisseur){
+double norme_coeff_reflexion(double theta_i, float eps_rel, float sigma, float epaisseur){
 
-    float theta_t = theta_t_snell(theta_i, eps_0, eps_rel*eps_0); //calcul angle de transmission
-    float s = epaisseur/cos(theta_t); // distance de propagation dans le mur (pour un "aller de rayon")
+    double theta_t = theta_t_snell(theta_i, eps_0, eps_rel*eps_0); //calcul angle de transmission
+    double s = epaisseur/cos(theta_t); // distance de propagation dans le mur (pour un "aller de rayon")
 
     float z1 = 377; //impedance caracteristique du vide
     float complex z3 = z2(eps_rel, sigma); //impedance caract du mur (COMPLEXE)
@@ -63,28 +63,28 @@ float norme_coeff_reflexion(float theta_i, float eps_rel, float sigma, float epa
 
     float complex gamma = gamma_m(eps_rel, sigma); //calcul de la constante de propagation (COMPLEX)
 
-    float complex coef_perp = (z3*cos(theta_i) - z1*cos(theta_t)) / (z3*cos(theta_i) + z1*cos(theta_t)); //coefficient de reflexion pour
+    double complex coef_perp = (z3*cos(theta_i) - z1*cos(theta_t)) / (z3*cos(theta_i) + z1*cos(theta_t)); //coefficient de reflexion pour
                                                                                                      //polarisation perpendiculaire
 
-    float complex partie_exp = cexp(-2*gamma*s) * cexp(I*beta*2*s*sin(theta_t)*sin(theta_i));  //Partie exponentielle de la formule du
+    double complex partie_exp = cexp(-2*gamma*s) * cexp(I*beta*2*s*sin(theta_t)*sin(theta_i));  //Partie exponentielle de la formule du
                                                                                                //coef de reflexion du mur
     return cabs( coef_perp + (1-pow(coef_perp,2))*(coef_perp*partie_exp/(1-pow(coef_perp,2)*partie_exp))  );
 }
 
-float norme_coeff_transmission(float theta_i, float eps_rel, float sigma, float epaisseur){
+double norme_coeff_transmission(double theta_i, float eps_rel, float sigma, float epaisseur){
 
-    float theta_t = theta_t_snell(theta_i, eps_0, eps_rel*eps_0); //calcul angle de transmission
-    float s = epaisseur/cos(theta_t); // distance de propagation dans le mur (pour un "aller de rayon")
+    double theta_t = theta_t_snell(theta_i, eps_0, eps_rel*eps_0); //calcul angle de transmission
+    double s = epaisseur/cos(theta_t); // distance de propagation dans le mur (pour un "aller de rayon")
 
     float z1 = 377; //impedance caracteristique du vide
     float complex z3 = z2(eps_rel, sigma); //impedance caract du mur (COMPLEXE)
 
     float complex gamma = gamma_m(eps_rel, sigma); //calcul de la constante de propagation (COMPLEX)
 
-    float complex coef_perp = (z3*cos(theta_i) - z1*cos(theta_t)) / (z3*cos(theta_i) + z1*cos(theta_t)); //coefficient de reflexion pour
+    double complex coef_perp = (z3*cos(theta_i) - z1*cos(theta_t)) / (z3*cos(theta_i) + z1*cos(theta_t)); //coefficient de reflexion pour
                                                                                                      //polarisation perpendiculaire
 
-    float complex partie_exp = cexp(-2*gamma*s) * cexp(I*beta*2*s*sin(theta_t)*sin(theta_i));  //Partie exponentielle de la formule du
+    double complex partie_exp = cexp(-2*gamma*s) * cexp(I*beta*2*s*sin(theta_t)*sin(theta_i));  //Partie exponentielle de la formule du
                                                                                             // du coef de reflexion du mur
 
     return cabs( ((1-pow(coef_perp,2))*cexp(-s*gamma)) / ((1-pow(coef_perp,2)*partie_exp)));
